[#50] Create get segment by case id endpoint
